+++ CHANGES.md
@@ -37,6 +37,7 @@ Release Notes.
 * Support collecting logs and forwarding through gRPC.
 * Support config `agent.sample_n_per_3_secs` can be changed in the runtime.
 * Support DNS periodic resolving mechanism to update backend service.
+* Support config `agent.trace.ignore_path` can be changed in the runtime.
 
 #### OAP-Backend
 * Make meter receiver support MAL.